+++ src/App.js
@@ -1,9 +1,8 @@
-import React from 'react';
 import { Switch, Route } from "react-router-dom";
 import { Container, Row, Col } from 'react-bootstrap';
 import Navigation from "./js/components/Navigation";
-import Accounts from "./js/components/connected/Accounts";
-import Footer from "./js/components/connected/Footer";
+import AccountsPresentational from "./js/components/presentationals/AccountsPresentational";
+import FooterPresentational from "./js/components/presentationals/FooterPresentational";
 
 function App() {
   return (
@@ -15,7 +14,7 @@ function App() {
             <Switch>
               <Route path="/accounts">
                 <h1>Accounts</h1>
-                <Accounts />
+                <AccountsPresentational />
               </Route>
               <Route path="/">
                 <h1>Home</h1>
@@ -23,8 +22,13 @@ function App() {
             </Switch>
           </Col>
         </Row>
+        <Row>
+          <Col>
+            <FooterPresentational />
+          </Col>
+        </Row>
       </Container>
-      <Footer />
+      
     </>
   );
 }