+++ src/js/__tests__/Footer.test.js
@@ -1,39 +1,45 @@
-import { LOADED_API_VERSION_SUCCESSFULLY } from "../constants/action-types";
-import { fetchBackendApiVersion } from "../actions/index";
 import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import nock from "nock";
+import { shallow, configure, mount } from "enzyme";
+import Adapter from "enzyme-adapter-react-16";
+import FooterPresentational from '../components/presentationals/FooterPresentational';
+import { render } from '@testing-library/react';
+import { actions as pingActions } from "../actions/PingActions";
 
-describe('Testing fetchBackendApiVersion()', () => {
+describe('Testing Footer Component', () => {
+
+  const payload = {
+    "app_version": "1.1.1"
+  };
+
+  nock(process.env.REACT_APP_BACKEND_URL || 'https://backend-flaskrestx.herokuapp.com/api/v1')
+  .get('/ping')
+  .reply(200,
+        payload,
+        {'Access-Control-Allow-Origin': '*'}
+  );
 
   const middlewares = [thunk];
   const mockStore = configureStore(middlewares);
-  const store = mockStore({});
+  const initialState = {
+    ping: {
+      appVersion: "2.2.2",
+      apiVersion: "1.1.1"
+    },
+    fetch: () => dispatch(pingActions.readAll())
+  }
+  const store = mockStore(initialState);
+
+  configure({adapter: new Adapter()})
 
   beforeEach(() => {
     store.clearActions();
   });
-  it('Should get LOADED_API_VERSION_SUCCESSFULLY', async () => {
-    const payload = {
-      "status": "success",
-      "message": "system up and running",
-      "api_version": "0.4.0"
-    };
-      
-    nock(process.env.REACT_APP_BACKEND_URL || 'https://backend-flaskrestx.herokuapp.com/api/v1')
-    .get('/ping')
-    .reply(200,
-      payload,
-      {'Access-Control-Allow-Origin': '*'}
-    );
 
-    let expectedActions = [{
-      type: LOADED_API_VERSION_SUCCESSFULLY,
-      payload: payload
-    }];
 
-    await store.dispatch(fetchBackendApiVersion()).then(() => {
-      expect(store.getActions()).toEqual(expectedActions);
-    });
+  it('Should renders footer text', () => {
+    const { getByText } = render(<FooterPresentational store={store} />);
+    expect(getByText('Backend Api Version: 1.1.1 - Frontend Version: 2.2.2')).toBeTruthy();
   });
 })