+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# 0.8.0
+# 0.9.0
++ Split reducers
++ Refactor store with typescript
++ Refactor footer
++ Change backend URL
+
+## 0.8.0
 + Add loading to fetch accounts list
 + Add Docker
 + Add Heroku deploy